decimal point totl ascres

+++ database/migrations/2024_10_15_094028_create_agriculture_data_table.php
@@ -19,7 +19,8 @@ return new class extends Migration
             $table->string('crop_variety'); // Nullable crop variety column
             $table->date('planting_date'); // Nullable planting date column
             $table->string('inputs'); // Nullable inputs column
-            $table->float('total_acres', 8, 2); // Integer total_acres
+            $table->decimal('total_acres', 8, 2);
+ // Integer total_acres
             $table->timestamps();
         });
